change square to uint8_t

+++ player/move_gen.h
@@ -23,7 +23,7 @@
 // board is 10 x 10
 #define BOARD_WIDTH 10
 
-typedef int square_t;
+typedef uint8_t square_t;
 typedef int8_t rnk_t;
 typedef int8_t fil_t;
